# api.yaml
---
openapi: 3.1.0
info:
  title: DatabaseManagerAPI
  version: 0.1.0
paths:
  "/get_configs/":
    get:
      summary: Read Resources
      operationId: read_resources_get_configs__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  "$ref": "#/components/schemas/ResourceConfig"
                type: array
                title: Response Read Resources Get Configs  Get
      security:
        - APIKeyHeader: [ ]
  "/add_new_config/":
    post:
      summary: Create Resource
      operationId: create_resource_add_new_config__post
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ResourceConfigCreate"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ResourceConfig"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      security:
        - APIKeyHeader: [ ]
  "/rm_config/{resource_id}":
    delete:
      summary: Delete Resource
      operationId: delete_resource_rm_config__resource_id__delete
      security:
        - APIKeyHeader: [ ]
      parameters:
        - name: resource_id
          in: path
          required: true
          schema:
            type: integer
            title: Resource Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ResourceConfig"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/":
    get:
      summary: Root
      operationId: root__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { }
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            "$ref": "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ResourceConfig:
      properties:
        name:
          type: string
          title: Name
        url:
          type: string
          title: Url
        destination:
          type: string
          title: Destination
        resource_id:
          type: integer
          title: Resource Id
      type: object
      required:
        - name
        - url
        - destination
        - resource_id
      title: ResourceConfig
    ResourceConfigCreate:
      properties:
        name:
          type: string
          title: Name
        url:
          type: string
          title: Url
        destination:
          type: string
          title: Destination
      type: object
      required:
        - name
        - url
        - destination
      title: ResourceConfigCreate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: access_token